chore: remove unused component
