--- src/pages/Vaults/Vault/components/TransactionHistory/index.tsx
@@ -1,2 +1,0 @@
-export type { TransactionHistoryProps } from './TransactionHistory';
-export { TransactionHistory } from './TransactionHistory';
