+++ src/pages/Vaults/Vault/components/index.tsx
@@ -1,17 +1,8 @@
 import { InsightListItem, InsightsList, InsightsListProps } from './InsightsList';
 import { PageTitle, PageTitleProps } from './PageTitle';
 import { Rewards, RewardsProps } from './Rewards';
-import { TransactionHistory, TransactionHistoryProps } from './TransactionHistory';
 import { VaultCollateral, VaultCollateralProps } from './VaultCollateral';
 import { VaultInfo, VaultInfoProps } from './VaultInfo';
 
-export { InsightsList, PageTitle, Rewards, TransactionHistory, VaultCollateral, VaultInfo };
-export type {
-  InsightListItem,
-  InsightsListProps,
-  PageTitleProps,
-  RewardsProps,
-  TransactionHistoryProps,
-  VaultCollateralProps,
-  VaultInfoProps
-};
+export { InsightsList, PageTitle, Rewards, VaultCollateral, VaultInfo };
+export type { InsightListItem, InsightsListProps, PageTitleProps, RewardsProps, VaultCollateralProps, VaultInfoProps };
